refactor(CarManager -> RacingGame): 도메인에 적합한 클래스 명으로 수정
