+++ __tests__/RacingGame.test.js
@@ -1,4 +1,4 @@
-const CarManager = require('../src/CarManager');
+const RacingGame = require('../src/RacingGame');
 
 const RandomNumberGenerator = require('../src/utils/RandomNumberGenerator');
 
@@ -10,22 +10,22 @@ const mockRandoms = (numbers) => {
 };
 
 describe('자동차를 관리하는 클래스 테스트', () => {
-  let carManager;
+  let racingGame;
 
   beforeEach(() => {
-    carManager = new CarManager(['pobi', 'crong', 'honux']);
+    racingGame = new RacingGame(['pobi', 'crong', 'honux']);
   });
 
   test('자동차 관리자 인스턴스를 생성한다.', () => {
-    expect(carManager.getCars().length).toEqual(3);
+    expect(racingGame.getCars().length).toEqual(3);
   });
 
   test('자동차 각각을 전진시킨다', () => {
     mockRandoms([5, 2, 1]);
 
-    carManager.progress();
+    racingGame.progress();
 
-    expect(carManager.getCars().map((singleCar) => singleCar.getProgressCount())).toEqual(
+    expect(racingGame.getCars().map((singleCar) => singleCar.getProgressCount())).toEqual(
       [1, 0, 0],
     );
   });
@@ -34,18 +34,18 @@ describe('자동차를 관리하는 클래스 테스트', () => {
     test('우승자가 한 명인 경우', () => {
       mockRandoms([5, 2, 1]);
 
-      carManager.progress();
+      racingGame.progress();
 
-      const winners = carManager.getWinners();
+      const winners = racingGame.getWinners();
       expect(winners.length === 1 && winners[0] === 'pobi').toBe(true);
     });
 
     test('우승자가 두 명 이상인 경우', () => {
       mockRandoms([6, 7, 2]);
 
-      carManager.progress();
+      racingGame.progress();
 
-      const winners = carManager.getWinners();
+      const winners = racingGame.getWinners();
       expect(winners).toEqual(['pobi', 'crong']);
     });
   });
